requirements.txt
gherkin-official==24.1.0
pyinstaller==5.13.2
PySimpleGUI==4.60.5